requirements.txt
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
distro==1.9.0
filelock==3.18.0
fsspec==2025.7.0
gitdb==4.0.12
GitPython==3.1.45
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.177.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.74.0
grpcio-status==1.71.2
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.33.5
hyperframe==6.1.0
idna==3.10
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
MarkupSafe==3.0.2
MouseInfo==0.1.3
mpmath==1.3.0
mss==10.0.0
narwhals==1.48.1
networkx==3.5
numpy==2.2.6
openai==1.97.1
opencv-python==4.12.0.88
packaging==25.0
pandas==2.3.1
pillow==11.3.0
portalocker==3.2.0
proto-plus==1.26.1
protobuf==5.29.5
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyAutoGUI==0.9.54
pydantic==2.11.7
pydantic_core==2.33.2
pydeck==0.9.1
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyparsing==3.2.3
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytweening==1.2.0
pytz==2025.2
# pywin32==311
PyYAML==6.0.2
qdrant-client==1.15.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
rpds-py==0.26.0
rsa==4.9.1
safetensors==0.5.3
scikit-learn==1.7.1
scipy==1.16.0
screeninfo==0.8.1
sentence-transformers==5.0.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
streamlit==1.47.0
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tokenizers==0.21.2
toml==0.10.2
torch==2.7.1
tornado==6.5.1
tqdm==4.67.1
transformers==4.53.3
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
watchdog==6.0.0
